+++ Libplanet/Net/Swarm.cs
@@ -241,6 +241,7 @@ namespace Libplanet.Net
 
         public void Dispose()
         {
+            _workerCancellationTokenSource?.Cancel();
             Transport.Dispose();
         }
 
@@ -256,7 +257,6 @@ namespace Libplanet.Net
             CancellationToken cancellationToken = default(CancellationToken)
         )
         {
-            _workerCancellationTokenSource?.Cancel();
             _logger.Debug($"Stopping {nameof(Swarm<T>)}...");
             using (await _runningMutex.LockAsync())
             {